requirements.dev.txt
flake8>=3.7.9
isort>=4.3.21
mypy==0.761
pytest>=5.3.2
pytest-runner>=5.2
requests-mock>=1.7.0
yapf>=0.28.0
